Update individual settings from json file
Previously the default settings (defined in the source) were simply overwritten by those in the json file. Settings that were defined in the defaults, but not in the file, got deleted, and caused the application to crash. This behaviour has changed now: the default values for settings are preserved if they are not defined in the json file.